Add sound effects to slash, bomb and level up
Added sound effects to the slash attack, the bomb attack and the bomb dot attack, and to when the player levels up.

Added an ambient sound effect of frogs. to the scene.

Changed the sorting layer of the bomb dot effect to 1 so that it will be "on the ground".